requirements.txt
Flask
Pillow
transparent_background 